+++ crucible-syntax/src/Lang/Crucible/Syntax/Atoms.hs
@@ -78,7 +78,7 @@ data Keyword = Defun | DefBlock | DefGlobal | Declare | Extern
              | Assert_ | Assume_
              | SetRegister
              | Funcall
-             | Breakpoint_
+             | Cut_
              | BV | BVConcat_ | BVSelect_ | BVTrunc_
              | BVZext_ | BVSext_ | BVNonzero_ | BoolToBV_
              | BVCarry_ | BVSCarry_ | BVSBorrow_
@@ -124,7 +124,7 @@ keywords =
   , ("assert!", Assert_)
   , ("assume!", Assume_)
   , ("funcall", Funcall)
-  , ("breakpoint", Breakpoint_)
+  , ("cut", Cut_)
 
     -- types
   , ("Any" , AnyT)