+++ crucible-llvm/CHANGELOG.md
@@ -1,3 +1,11 @@
+# next
+
+* **BREAKING:** Rename various bits associated with the "breakpoint"
+  feature in accordance with renaming the feature to "cutpoint".
+  In particular, `testBreakpointFunction` is now `testCutpointFunction`.
+  Also, the family of LLVM symbols recognized now begins with `__cutpoint__`
+  rather than `__breakpoint__`.
+
 # 0.9 -- 2026-01-29
 
 * The `LLVM_Debug` data constructor for `LLVMStmt`, as well as the related